+++ optimacore/system.py
@@ -39,12 +39,6 @@ class SystemSettings(object):
     DEFAULT_SYMBOL_YES = "y"
     DEFAULT_SYMBOL_NO = "n"
     
-    EXCEL_IO_DEFAULT_COMMENT_XSCALE = 3
-    EXCEL_IO_DEFAULT_COMMENT_YSCALE = 3
-    EXCEL_IO_DEFAULT_COLUMN_WIDTH = 20
-    EXCEL_FILE_EXTENSION = ".xlsx"
-    EXCEL_LIST_SEPARATOR = ","
-    
 
 #%% Code for determining module installation directory.